+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 You can find the changelog in this repository's Github releases page: <https://github.com/grafana/terraform-provider-grafana/releases>.
+
+## Unreleased
+
+- Bump Grafana App SDK to v0.50.1 and Grafana Plugin SDK Go to v0.287.0.